move the net message code to its own Python file
